consolidate odds and ends into general bits module
