+++ ui/bits/package.json
@@ -46,45 +46,34 @@
   "lichess": {
     "bundles": [
       "src/bits.account.ts",
-      "src/bits.appeal.ts",
-      "src/bits.autoform.ts",
       "src/bits.captcha.ts",
       "src/bits.challengePage.ts",
       "src/bits.checkout.ts",
       "src/bits.clas.ts",
       "src/bits.cms.ts",
       "src/bits.coachForm.ts",
-      "src/bits.colorizeYesNoTable.ts",
       "src/bits.confetti.ts",
-      "src/bits.contact.ts",
       "src/bits.cropDialog.ts",
-      "src/bits.dailyFeed.ts",
       "src/bits.devMode.ts",
       "src/bits.diagnosticDialog.ts",
-      "src/bits.eventCountdown.ts",
       "src/bits.expandText.ts",
       "src/bits.flairPicker.ts",
       "src/bits.flatpickr.ts",
       "src/bits.forum.ts",
       "src/bits.gameSearch.ts",
-      "src/bits.hcaptcha.ts",
-      "src/bits.importer.ts",
       "src/bits.infiniteScroll.ts",
       "src/bits.login.ts",
       "src/bits.lpv.ts",
-      "src/bits.oauth.ts",
       "src/bits.passwordComplexity.ts",
       "src/bits.plan.ts",
       "src/bits.polyglot.ts",
       "src/bits.publicChats.ts",
       "src/bits.qrcode.ts",
-      "src/bits.relayForm.ts",
       "src/bits.soundMove.ts",
-      "src/bits.streamer.ts",
       "src/bits.team.ts",
       "src/bits.teamBattleForm.ts",
-      "src/bits.titleRequest.ts",
       "src/bits.tourForm.ts",
+      "src/bits.ts",
       "src/bits.tvGames.ts",
       "src/bits.ublog.ts",
       "src/bits.ublogForm.ts",
