--- ui/bits/src/bits.contact.ts
@@ -1,6 +1,0 @@
-import contactEmail from './bits.contactEmail';
-
-site.load.then(() => {
-  location.hash ||= '#help-root';
-  contactEmail();
-});
