--- ui/bits/src/bits.dailyFeed.ts
@@ -1,5 +1,0 @@
-import flairPickerLoader from './exports/flairPicker';
-
-$('.emoji-details').each(function(this: HTMLElement) {
-  flairPickerLoader(this);
-});
