--- ui/bits/src/bits.importer.ts
@@ -1,14 +1,0 @@
-import { spinnerHtml } from 'common/spinner';
-
-const $form: Cash = $('main.importer form');
-
-$form.on('submit', () => setTimeout(() => $form.html(spinnerHtml), 50));
-
-$form.find('input[type=file]').on('change', function(this: HTMLInputElement) {
-  const file = this.files?.[0];
-  if (!file) return;
-
-  const reader = new FileReader();
-  reader.onload = e => $form.find('textarea').val(e.target?.result as string);
-  reader.readAsText(file);
-});
